pyproject.toml
[project]
name = "ComfyUI-DMXAPI"
description = "DMXAPI integration for ComfyUI with Seedream-3.0 text-to-image model"
version = "1.0.0"
license = {file = "LICENSE"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.28.0",
    "Pillow>=9.0.0",
    "numpy>=1.22.0",
    "torch>=1.12.0"
]

[project.urls]
Homepage = "https://www.dmxapi.com"
Repository = "https://github.com/yourusername/ComfyUI-DMXAPI"
Documentation = "https://www.dmxapi.com/docs"
Issues = "https://github.com/yourusername/ComfyUI-DMXAPI/issues"

[tool.comfy]
PublisherId = "your-publisher-id"  # 替换为您的ComfyUI发布者ID
DisplayName = "DMXAPI Nodes"
Icon = "icon.png"  # 可选，插件图标
Tags = ["text-to-image", "ai-generation", "seedream-3.0"]
Compatibility = ">=1.0.0"

[build-system]
requires = ["setuptools>=61.0.0"]
build-backend = "setuptools.build_meta"